Moved grid to util
