+++ src/components/banlist/BanListSection.js
@@ -6,7 +6,7 @@ import { Typography, Box } from '@material-ui/core'
 
 import { ChildBox } from '../MainContent'
 
-import CardDisplayGrid from '../grid/CardDisplayGrid'
+import CardDisplayGrid from '../util/grid/CardDisplayGrid'
 
 
 const BanListSection = ( { sectionExplanation, cards, isDataLoaded } ) =>
@@ -42,7 +42,7 @@ const BanListSection = ( { sectionExplanation, cards, isDataLoaded } ) =>
 
 	return (
 		<ChildBox >
-			<Box style={{backgroundColor: '#f6f2fb', padding: '1.4rem', margin: '1.5rem', borderRadius: '1.25rem'}} >
+			<Box style={{backgroundColor: '#f6f2fb', padding: '1.4rem', maxWidth: '85%', borderRadius: '1.25rem', margin: 'auto', marginTop: '2rem', marginBottom: '2rem'}} >
 				<SectionInfoText variant='subtitle2' align='center' >
 					{ sectionExplanation }
 				</SectionInfoText>
